refactor: extract bac calculator and drink modal into separate bac module
