+++ app/screens/profile/profileTypes.ts
@@ -1,0 +1,11 @@
+import type { DrinkEntry } from '../../types/drinkTypes';
+
+export type ProfileUserInfo = {
+  name?: string;
+  username?: string;
+  weight?: number;
+  gender?: 'male' | 'female';
+  drinks?: DrinkEntry[];
+  bacHighscoreAllTime?: number;
+  bacHighscoreUpdatedAt?: number;
+};
